last commit

+++ database/migrations/2023_04_07_121320_create_concour_user_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
              
-            $table->integer('note');
+            $table->float('note');
             $table->timestamps();
 
         });
